Move CompiledBlob to a new file
